Rename to gocheckcompilerdirectives
