+++ cmd/gocheckcompilerdirectives/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/leighmcculloch/gocheckcompilerdirectives/checkcompilerdirectives"
+	"golang.org/x/tools/go/analysis/singlechecker"
+)
+
+func main() {
+	singlechecker.Main(checkcompilerdirectives.Analyzer())
+}
